Routines to obtain number of cycles ramped through and tests.
However, the number of cycles calculation isn't quite that
accurate, maybe only to within 0.001 - 0.01 cycles
(0.006 - 0.06 radians), which maybe isn't very good,
particularly for fluctuations...

The accuracy (or lack thereof) doesn't seem to be fully
attributable to errors in the determination of the signal
frequency... so where is it coming from? Can we improve it?
Maybe some sort of "indexing" error? Or perhaps we can
improve with a cubic (as opposed to linear) fit?