Cleaned up sine fitting, allowing user to specify # fitting pts.
Although arbitrary fitting points are allowed (well, technically,
we only support *even* numbers of fitting points such that the
number of points above and below the zero crossing are equal),
in practice, for the types of data I'm interested in,
in which there are perhaps ~10 points per cycle,

                      Npts = 4

appeared to work optimally. Npts <= 3 is too few to constrain
a half cycle of a sinusoid, and Npts >= 6 appears to overconstrain
the problem and gives erroneous results (?).