Copy signal and its slices to avoid altering original copy.
The previous commit mentioned that `Npts` was not effective
for `Npts` >= 6; however, this was due to an error resulting
from the modification of the signal because we were not
copying the signal slices and subsequently modifying them.
The sinusoidal fitting works well for large `Npts`
(even up to ~100), but no significant or systematic
improvement has been seen for `Npts` > 4. Thus, it is
still probably best to use `Npts` = 4.