Added finishing touches on decimate(...) and corresponding tests.
In particular, corrected off-by-one error in start index of
decimated `valid` slice.

Strange errors regarding a TypeError when `t` is not specified
for *some* of the test cases... don't fully understand it.
Although, by running PDB and causing decimate(...) to err,
I was manually able to confirm that type(t) was None,
as it should have been specified per the defaults,
despite nosetests giving me the TypeError.
Something strange is going on with the nosetests that is
deeper than I can understand, so just explicitly passing
the argument for `t` works for now...

Also, the decimate(...) test with a realistic signal reduces
high frequency noise by ~1 order of magnitude. I was kind of
hoping for more, though...